# swagger.json
{
    "swagger": "2.0",
    "info": {
      "title": "Crisis management calculations",
      "description": "API list for calculation of cross sections, load and joint detais of shelter.",
      "version": "1.0"
    },
    "basePath": "",
    "schemes": [
      "http"
    ],
    "consumes": [
      "application/json"
    ],
    "produces": [
      "application/json"
    ],
    "paths": {
      "/cross_section": {
        "post": {
          "summary": "Calculate cross section values based on input attributes.",
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "in": "body",
              "name": "body",
              "description": "Attributes for material, load, and footprint",
              "required": true,
              "schema": {
                "type": "object",
                "properties": {
                  "material": {
                    "type": "object",
                    "properties": {
                      "partial_factor": { 
                        "type": "number",
                        "example": 1.3
                       },
                      "density": { 
                        "type": "number",
                        "example": 610
                       },
                      "bending_strength": { 
                        "type": "number",
                        "example": 91
                       },
                      "shear_strength": { 
                        "type": "number",
                        "example": 9
                       },
                      "compression_parallel": { 
                        "type": "number",
                        "example": 45
                       },
                      "e_modulus": { 
                        "type": "number",
                        "example": 11.2
                       },
                      "e_modulus_5": { 
                        "type": "number",
                        "example": 11.2
                       },
                      "modification_factor_permanent_term": { 
                        "type": "number",
                        "example": 0.5
                       },
                      "modification_factor_medium_term": { 
                        "type": "number",
                        "example": 0.65
                       },
                      "modification_factor_instantaneous_term": { 
                        "type": "number",
                        "example": 0.9
                       },
                      "creep_factor": { 
                        "type": "number",
                        "example": 2
                       },
                      "creep_factor_solid_timber": { 
                        "type": "number",
                        "example": 0.2
                       }
                    }
                  },
                  "load": {
                    "type": "object",
                    "properties": {
                      "P_L": { 
                        "type": "number",
                        "example": 2.96
                       },
                      "M_L": { 
                        "type": "number",
                        "example": 7.44
                       },
                      "I_L": { 
                        "type": "number",
                        "example": 5.66
                       },
                      "SLS_L": { 
                        "type": "number",
                        "example": 5.18
                       },
                      "gk": { 
                        "type": "number",
                        "example": 1.38
                       },
                      "g_lead": { 
                        "type": "number",
                        "example": 2
                       },
                      "g_acmp": { 
                        "type": "number",
                        "example": 1.4
                       },
                      "psi_lead": { 
                        "type": "number",
                        "example": 0
                       },
                      "psi_acmp": { 
                        "type": "number",
                        "example": 0.2
                       },
                      "P_clm": { 
                        "type": "number",
                        "example": 3.7
                       },
                      "M_clm": { 
                        "type": "number",
                        "example": 9.2
                       },
                      "I_clm": { 
                        "type": "number",
                        "example": 7
                       }
                    }
                  },
                  "footprint": {
                    "type": "object",
                    "properties": {
                      "beam_length": { 
                        "type": "number",
                        "example": 2
                       },
                      "height": { 
                        "type": "number",
                        "example": 2
                       }
                    }
                  }
                }
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Successful response with cross-section values.",
              "schema": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "weight": {
                      "type": "object",
                      "properties": {
                        "print_value": {
                          "type": "string",
                          "example": "4.36 kg"
                        },
                        "value": {
                          "type": "number",
                          "example": 4.36
                        }
                      }
                    },
                    "width": {
                      "type": "object",
                      "properties": {
                        "print_value": {
                          "type": "string",
                          "example": "65 mm"
                        },
                        "value": {
                          "type": "number",
                          "example": 65
                        }
                      }
                    },
                    "thickness": {
                      "type": "object",
                      "properties": {
                        "print_value": {
                          "type": "string",
                          "example": "55 mm"
                        },
                        "value": {
                          "type": "number",
                          "example": 55
                        }
                      }
                    },
                    "length": {
                      "type": "object",
                      "properties": {
                        "print_value": {
                          "type": "string",
                          "example": "2.00 m"
                        },
                        "value": {
                          "type": "number",
                          "example": 2.0
                        }
                      }
                    },
                    "bending_utilisation": {
                      "type": "object",
                      "properties": {
                        "print_value": {
                          "type": "string",
                          "example": "249.48%"
                        },
                        "value": {
                          "type": "number",
                          "example": 249.48
                        }
                      }
                    },
                    "bending_status": {
                      "type": "object",
                      "properties": {
                        "print_value": {
                          "type": "string",
                          "example": "Unacceptable"
                        },
                        "value": {
                          "type": "boolean",
                          "example": false
                        }
                      }
                    },
                    "shear_utilisation": {
                      "type": "object",
                      "properties": {
                        "print_value": {
                          "type": "string",
                          "example": "69.37%"
                        },
                        "value": {
                          "type": "number",
                          "example": 69.37
                        }
                      }
                    },
                    "shear_status": {
                      "type": "object",
                      "properties": {
                        "print_value": {
                          "type": "string",
                          "example": "Acceptable"
                        },
                        "value": {
                          "type": "boolean",
                          "example": true
                        }
                      }
                    },
                    "sls_utilisation": {
                      "type": "object",
                      "properties": {
                        "print_value": {
                          "type": "string",
                          "example": "1603.77%"
                        },
                        "value": {
                          "type": "number",
                          "example": 1603.77
                        }
                      }
                    },
                    "sls_status": {
                      "type": "object",
                      "properties": {
                        "print_value": {
                          "type": "string",
                          "example": "Unacceptable"
                        },
                        "value": {
                          "type": "boolean",
                          "example": false
                        }
                      }
                    },
                    "compression_utilisation": {
                      "type": "object",
                      "properties": {
                        "print_value": {
                          "type": "string",
                          "example": "11.44%"
                        },
                        "value": {
                          "type": "number",
                          "example": 11.44
                        }
                      }
                    },
                    "compression_status": {
                      "type": "object",
                      "properties": {
                        "print_value": {
                          "type": "string",
                          "example": "Acceptable"
                        },
                        "value": {
                          "type": "boolean",
                          "example": true
                        }
                      }
                    },
                    "buckling_utilisation_in_plane": {
                      "type": "object",
                      "properties": {
                        "print_value": {
                          "type": "string",
                          "example": "32.12%"
                        },
                        "value": {
                          "type": "number",
                          "example": 32.12
                        }
                      }
                    },
                    "buckling_status_in_plane": {
                      "type": "object",
                      "properties": {
                        "print_value": {
                          "type": "string",
                          "example": "Acceptable"
                        },
                        "value": {
                          "type": "boolean",
                          "example": true
                        }
                      }
                    },
                    "buckling_utilisation_out_of_plane": {
                      "type": "object",
                      "properties": {
                        "print_value": {
                          "type": "string",
                          "example": "23.39%"
                        },
                        "value": {
                          "type": "number",
                          "example": 23.39
                        }
                      }
                    },
                    "buckling_status_out_of_plane": {
                      "type": "object",
                      "properties": {
                        "print_value": {
                          "type": "string",
                          "example": "Acceptable"
                        },
                        "value": {
                          "type": "boolean",
                          "example": true
                        }
                      }
                    },
                    "final_utilisation": {
                      "type": "object",
                      "properties": {
                        "print_value": {
                          "type": "string",
                          "example": "1603.77%"
                        },
                        "value": {
                          "type": "number",
                          "example": 1603.77
                        }
                      }
                    },
                    "final_utilisation_status": {
                      "type": "object",
                      "properties": {
                        "print_value": {
                          "type": "string",
                          "example": "Unacceptable"
                        },
                        "value": {
                          "type": "boolean",
                          "example": false
                        }
                      }
                    }
                  }
                }
              }
            },
            "400": {
              "description": "Invalid request parameters."
            }
          }
        }
      },
...
      "/joint3": {
        "post": {
          "summary": "Optimizes the joint3 details based on provided material and footprint attributes.",
          "produces": [
            "application/json"
          ],
          "consumes": [
            "application/json"
          ],
          "parameters": [
            {
              "in": "body",
              "name": "body",
              "description": "Material and footprint attributes for joint3 details optimization.",
              "required": true,
              "schema": {
                "type": "object",
                "properties": {
                  "cross_section": {
                    "type": "object",
                    "properties": {
                      "column_h": {
                        "type": "number",
                        "example": 100
                      },
                      "column_w": {
                        "type": "number",
                        "example": 112.5
                      },
                      "tie_beam_h": {
                        "type": "number",
                        "example": 152.4
                      }
                    }
                  },
                  "material": {
                    "type": "object",
                    "properties": {
                      "dtl_e": {
                        "type": "number",
                        "example": 2
                      },
                      "dtl_s": {
                        "type": "number",
                        "example": 2.5
                      },
                      "dtl_v": {
                        "type": "number",
                        "example": 1.5
                      },
                      "dtl_g": {
                        "type": "number",
                        "example": 1.5
                      }
                    }
                  }
                }
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Successful response",
              "schema": {
                "type": "object",
                "properties": {
                  "capacity_kN": {
                    "type": "object",
                    "properties": {
                      "print_value": {
                        "type": "string",
                        "example": "11.04 kN"
                      },
                      "value": {
                        "type": "number",
                        "example": 11.04
                      }
                    }
                  },
                  "d": {
                    "type": "object",
                    "properties": {
                      "print_value": {
                        "type": "string",
                        "example": "25.91 mm"
                      },
                      "value": {
                        "type": "number",
                        "example": 25.91
                      }
                    }
                  },
                  "lim_e": {
                    "type": "object",
                    "properties": {
                      "print_value": {
                        "type": "string",
                        "example": "54.69 mm"
                      },
                      "value": {
                        "type": "number",
                        "example": 54.69
                      }
                    }
                  },
                  "lim_g": {
                    "type": "object",
                    "properties": {
                      "print_value": {
                        "type": "string",
                        "example": "41.02 mm"
                      },
                      "value": {
                        "type": "number",
                        "example": 41.02
                      }
                    }
                  },
                  "lim_s": {
                    "type": "object",
                    "properties": {
                      "print_value": {
                        "type": "string",
                        "example": "68.36 mm"
                      },
                      "value": {
                        "type": "number",
                        "example": 68.36
                      }
                    }
                  },
                  "lim_v": {
                    "type": "object",
                    "properties": {
                      "print_value": {
                        "type": "string",
                        "example": "41.02 mm"
                      },
                      "value": {
                        "type": "number",
                        "example": 41.02
                      }
                    }
                  },
                  "status": {
                    "type": "object",
                    "properties": {
                      "print_value": {
                        "type": "string",
                        "example": "Acceptable"
                      },
                      "value": {
                        "type": "boolean",
                        "example": true
                      }
                    }
                  }
                }
              }
            },
            "400": {
              "description": "Invalid input provided."
            }
          }
        }
      },
      "/generate_g_code" : {
        "post": {
          "summary": "Gets some cross section values and returns g code for cnc",
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "in": "body",
              "name": "body",
              "description": "Cross section for generate gcode of cnc",
              "required": true,
              "schema": {
                "type": "object",
                "properties": {
                  "tie_beam_w": {
                    "type": "number",
                    "description": "Tie beam width",
                    "example": 20
                  },
                  "tie_beam_h": {
                    "type": "number",
                    "description": "Tie beam height",
                    "example": 40
                  },
                  "column_h": {
                    "type": "number",
                    "description": "Column height",
                    "example": 10
                  }
                }
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Successful response with G-code commands as a string",
              "schema": {
                "type": "string",
                "example": "G21\nG17\nG90\nF1000\nS10000\nG0 Z10\nG0 X4.0 Y4.0\nG1 Z-2\nG1 X16.0\nG1 Y36.0\nG1 X4.0\nG1 Y4.0\nG0 Z10\n..."
              }
            },
            "400": {
              "description": "Invalid input"
            }
          }
        }
      },
      "/design_verify" : {
        "post": {
          "summary": "Verify if design is acceptable or not.",
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "in": "body",
              "name": "body",
              "description": "gets array of designs and return just the acceptable designs id",
              "required": true,
              "schema": {
                "type": "object",
                "properties": {
                  "designs": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "number",
                          "example": 1
                        },
                        "material": {
                          "type": "object",
                          "properties": {
                            "partial_factor": {
                              "type": "number",
                              "example": 1.3
                            },
                            "density": {
                              "type": "number",
                              "example": 610
                            },
                            "bending_strength": {
                              "type": "number",
                              "example": 91
                            },
                            "shear_strength": {
                              "type": "number",
                              "example": 9
                            },
                            "compression_parallel": {
                              "type": "number",
                              "example": 45
                            },
                            "e_modulus": {
                              "type": "number",
                              "example": 11.2
                            },
                            "e_modulus_5": {
                              "type": "number",
                              "example": 11.2
                            },
                            "modification_factor_permanent_term": {
                              "type": "number",
                              "example": 0.5
                            },
                            "modification_factor_medium_term": {
                              "type": "number",
                              "example": 0.65
                            },
                            "modification_factor_instantaneous_term": {
                              "type": "number",
                              "example": 0.9
                            },
                            "creep_factor": {
                              "type": "number",
                              "example": 2
                            },
                            "creep_factor_solid_timber": {
                              "type": "number",
                              "example": 0.2
                            }
                          }
                        },
                        "cross_section": {
                          "type": "object",
                          "properties": {
                            "beam_w": {
                              "type": "number",
                              "example": 0.1,
                              "description": "Width of the beam in meters"
                            },
                            "beam_h": {
                              "type": "number",
                              "example": 0.15,
                              "description": "Height of the beam in meters"
                            }
                          }
                        },
                        "footprint": {
                          "type": "object",
                          "properties": {
                            "length": {
                              "type": "number",
                              "example": 5.0
                            },
                            "column_number": {
                              "type": "integer",
                              "example": 4
                            },
                            "height": {
                              "type": "number",
                              "example": 3.0,
                              "description": "Height of the column in meters"
                            },
                            "slab_thickness": {
                              "type": "number",
                              "example": 0.2
                            },
                            "width": {
                              "type": "number",
                              "example": 2
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Successful response with acceptable designs.",
              "schema": {
                "type": "object",
                "properties": {
                  "acceptable_designs": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "description": "desgin object that user send it in input"
                    }
                  }
                }
              }
            },
            "400": {
              "description": "Invalid input"
            }
          }
        }
      }
    }    
  }